+++ packages/autocomplete/src/lib/util.ts
@@ -1,2 +1,0 @@
-// Mark a field as optional in an already defined type
-export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;