+++ packages/favorites/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-favorites@0.1.1...@availity/mui-favorites@0.1.2) (2024-08-01)
+
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-favorites@0.1.0...@availity/mui-favorites@0.1.1) (2024-07-22)
 
 ### Dependency Updates