+++ packages/table/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.15](https://github.com/Availity/element/compare/@availity/mui-table@0.1.14...@availity/mui-table@0.1.15) (2024-08-01)
+
+### Dependency Updates
+
+* `mui-chip` updated to version `0.1.14`
+* `mui-form-utils` updated to version `0.1.14`
+* `mui-textfield` updated to version `0.1.14`
 ## [0.1.14](https://github.com/Availity/element/compare/@availity/mui-table@0.1.13...@availity/mui-table@0.1.14) (2024-07-22)
 
 ### Dependency Updates